fix DbtInternalError import + move ConnectionError to common
