+++ core/dbt/common/exceptions/__init__.py
@@ -2,3 +2,4 @@ from dbt.common.exceptions.base import *  # noqa
 from dbt.common.exceptions.events import *  # noqa
 from dbt.common.exceptions.macros import *  # noqa
 from dbt.common.exceptions.contracts import *  # noqa
+from dbt.common.exceptions.connection import *  # noqa
