+++ core/dbt/common/exceptions/connection.py
@@ -1,0 +1,7 @@
+class ConnectionError(Exception):
+    """
+    There was a problem with the connection that returned a bad response,
+    timed out, or resulted in a file that is corrupt.
+    """
+
+    pass
