+++ tests/unit/test_registry_get_request_exception.py
@@ -1,6 +1,6 @@
 import unittest
 
-from dbt.exceptions import ConnectionError
+from dbt.common.exceptions import ConnectionError
 from dbt.clients.registry import _get_with_retries
